% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reveal_by_group.R
\name{reveal_by_group}
\alias{reveal_by_group}
\title{Reveal plot by group}
\usage{
reveal_by_group(p)
}
\arguments{
\item{p}{A ggplot2 object, in which a \code{group} aesthetic is used.}
}
\value{
A list of ggplot2 objects, which can be passed to \code{\link[=reveal_save]{reveal_save()}}
}
\description{
Creates a list of plots, showing data incrementally by groups.
}
\examples{
# Create full plot
library(ggplot2)
data("mtcars")

p <- mtcars |>
  ggplot(aes(mpg, wt,
             color = factor(vs),
             group = factor(vs))) +
  geom_point() +
  geom_smooth(method="lm",
              formula = 'y ~ x',
              linewidth=1) +
  facet_wrap(~am)
p

plot_list <- reveal_by_group(p)
plot_list[[1]]
plot_list[[2]]
plot_list[[3]]

\dontrun{
# Save plots
reveal_save(plot_list, "myplot", width = 8, height = 4)
}
}
