% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reveal_by_panel.R
\name{reveal_by_panel}
\alias{reveal_by_panel}
\title{Reveal plot by facet}
\usage{
reveal_by_panel(p, what = c("data", "everything"))
}
\arguments{
\item{p}{A ggplot2 object, in which a \code{group} aesthetic is used.}

\item{what}{one of \code{"data"} (the default), \code{"axis"}, \code{"label"} or \code{"everything"}.
With \code{"data"}, the basic graph layout (including axes and facet labels) is shown
from the start, and only the data points are shown incrementally. With \code{"axis"} and \code{"label"},
both the data and the corresponding axes and facet labels, respectively, are shown
incrementally. With \code{"everything"}, the entire panels are shown incrementally.}
}
\value{
A list of ggplot2 objects, which can be passed to \code{\link[=reveal_save]{reveal_save()}}
}
\description{
Creates a list of plots, showing data incrementally by layers.
}
\examples{
# Create full plot
library(ggplot2)
data("mtcars")

p <- mtcars |>
  ggplot(aes(mpg, wt,
             color = factor(vs),
             group = factor(vs))) +
  geom_point() +
  geom_smooth(method="lm",
              formula = 'y ~ x',
              linewidth=1) +
  facet_wrap(~am)
p

# Only data
plot_list <- reveal_by_panel(p, "data")
plot_list[[1]]
plot_list[[2]]
plot_list[[3]]

# Everything
plot_list <- reveal_by_panel(p, "everything")
plot_list[[1]]
plot_list[[2]]
plot_list[[3]]

\dontrun{
# Save plots
reveal_save(plot_list, "myplot", width = 8, height = 4)
}
}
