% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reveal_axis.R
\name{reveal_axis}
\alias{reveal_axis}
\alias{reveal_x}
\alias{reveal_y}
\title{Reveal plot by axis}
\usage{
reveal_axis(p, order = NULL, which = c("x", "y"))

reveal_x(p, order = NULL)

reveal_y(p, order = NULL)
}
\arguments{
\item{p}{A ggplot2 object}

\item{order}{(optional) A numeric vector specifying in which order to reveal the groups.}

\item{which}{(optional) "x" or "y", to specify the axis to be revealed. "x" is the default.}
}
\value{
A list of ggplot2 objects, which can be passed to \code{\link[=reveal_save]{reveal_save()}}
}
\description{
Creates a list of plots, showing data incrementally by axis (x or y) values. The specified axis must
must be mapped to a discrete variable. \code{reveal_x()} and \code{reveal_y()} are wrappers that call
\code{reveal_axis()} for the respective axes.
}
\examples{
# Create full plot
library(ggplot2)
data("mtcars")

p <- mtcars |>
  ggplot(aes(factor(vs), 
             color = gear,
             fill=ger, 
             group = gear)) +
  geom_bar() +
  facet_wrap(~am)
p

plot_list <- reveal_x(p)
plot_list[[1]]
plot_list[[2]]
plot_list[[3]]

\dontrun{
# Save plots
reveal_save(plot_list, "myplot", width = 8, height = 4)
}
}
