% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reveal_panels.R
\name{reveal_panels}
\alias{reveal_panels}
\title{Reveal plot by panel}
\usage{
reveal_panels(p, order = NULL, what = c("data", "everything"))
}
\arguments{
\item{p}{A ggplot2 object}

\item{order}{(optional) A numeric vector specifying in which order to reveal
the panels}

\item{what}{(optional) one of \code{"data"} or \code{"everything"}.' With \code{"data"}, the
basic graph layout (including axes and facet labels) is shown from the start,
and only the data points are shown incrementally. With \code{"everything"}, the
entire panels are shown incrementally.}
}
\value{
A list of ggplot2 objects, which can be passed to \code{\link[=reveal_save]{reveal_save()}}
}
\description{
Creates a list of plots, showing data incrementally by panels.
}
\examples{
# Create full plot
library(ggplot2)
data("mtcars")

p <- mtcars |>
  ggplot(aes(mpg, wt,
             color = factor(vs),
             group = factor(vs))) +
  geom_point() +
  geom_smooth(method="lm",
              formula = 'y ~ x',
              linewidth=1) +
  facet_wrap(~am)
p

# Only data
plot_list <- reveal_panels(p, what = "data")
plot_list[[1]]
plot_list[[2]]
plot_list[[3]]

# Everything
plot_list <- reveal_panels(p, what = "everything")
plot_list[[1]]
plot_list[[2]]
plot_list[[3]]

\dontrun{
# Save plots
reveal_save(plot_list, "myplot.png", width = 8, height = 4)
}
}
