% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reveal_patchwork.R
\name{reveal_patchwork}
\alias{reveal_patchwork}
\title{Reveal a patchwork}
\usage{
reveal_patchwork(p)
}
\arguments{
\item{p}{A patchwork object}
}
\value{
A list of ggplot2 objects, which can be passed to \code{\link[=reveal_save]{reveal_save()}}
}
\description{
Turns a \href{https://patchwork.data-imaginist.com/}{patchwork} into a list of plots that can be shown incrementally.
}
\examples{
# Compose plots with patchwork
if (require("patchwork")) {

   library(ggplot2)
   library(patchwork)
   data("mtcars")
   
   p1 <- ggplot(mtcars) +
           geom_point(aes(mpg, disp)) +
           ggtitle('Plot 1')
   
   p2 <- ggplot(mtcars) +
           geom_boxplot(aes(gear, disp, group = gear)) +
           ggtitle('Plot 2')
   
   p3 <- ggplot(mtcars) +
          geom_point(aes(hp, wt)) +
          ggtitle('Plot 3')
   
   p <- (p1 | p2) / p3
   p <- p + plot_annotation(title = 'The surprising truth about mtcars',
                            tag_levels = 'A'
                            )
   
   plot_list <- reveal_patchwork(p)
   plot_list[[1]]
   plot_list[[2]]
   plot_list[[3]]

}
}
