% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reveal_patchwork.R
\name{reveal_patchwork}
\alias{reveal_patchwork}
\title{Reveal plots in a patchwork object}
\usage{
reveal_patchwork(p, order = NULL)
}
\description{
Turns a patchwork into a list of plots that reveal each child plot (including
nested patchworks) one-by-one. The function handles arbitrary nesting levels.
}
