% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reveal_save.R
\name{reveal_save}
\alias{reveal_save}
\title{Saves incremental plots}
\usage{
reveal_save(plot_list, basename = "plot", ...)
}
\arguments{
\item{plot_list}{A list of plots created by \code{\link[=reveal_by_group]{reveal_by_group()}}, \code{\link[=reveal_by_layer]{reveal_by_layer()}} or \code{\link[=reveal_by_panel]{reveal_by_panel()}}}

\item{basename}{The base file name that will be used for saving}

\item{...}{Additional arguments (e.g. width, height) to be passed to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}}
}
\description{
Saves incremental plots
}
\examples{
# Create full plot
library(ggplot2)
data("mtcars")

p <- mtcars |>
  ggplot(aes(mpg, wt,
             color = factor(vs),
             group = factor(vs))) +
  geom_point() +
  geom_smooth(method="lm",
              formula = 'y ~ x',
              linewidth=1) +
  facet_wrap(~am)
p

plot_list <- reveal_by_group(p)
plot_list[[1]]
plot_list[[2]]
plot_list[[3]]

\dontrun{
# Save plots
reveal_save(plot_list, "myplot", width = 8, height = 4)
}
}
